requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.9.3
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
distlib==0.3.8
Django==5.1.1
django-extensions==3.2.3
djangorestframework==3.15.2
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.15.4
frappe==0.0.1
frappe-bench==5.22.9
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
honcho==1.1.0
idna==2.10
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.0
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pip-system-certs==4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.1.1
redis==5.0.8
requests==2.32.3
selenium==4.23.1
semantic-version==2.10.0
setuptools==72.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
supervisor==4.2.5
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
urllib3==1.26.19
urlopen==1.0.0
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0
wrapt==1.16.0
wsproto==1.2.0
